+++ web/src/app/app_view.dart
@@ -34,24 +34,16 @@ class AppView extends ui.Composite implements AppPresenterDisplay {
     topPanel.addStyleName("topPanel");
     topPanel.setSize("100%", "100%");
 
-    List<String> styles = ["clear", "chrome", "dark", "standard"];
-    // Add Style chooser
-    ui.ListBox styleChooser = new ui.ListBox();
-    styles.forEach((String style){
-      styleChooser.addItem(style, "packages/dart_web_toolkit/resource/${style}/${style}.css");
-    });
-    styleChooser.addChangeHandler(new event.ChangeHandlerAdapter((event.ChangeEvent evt){
-      int index = styleChooser.getSelectedIndex();
-      String item = styleChooser.getItemText(index);
-      String value = styleChooser.getValue(index);
-      print("Selected ${item} : ${value}");
-    }));
+    StyleChooserComponent styleChooser = new StyleChooserComponent("Style:");
     topPanel.addEast(styleChooser, 200.0);
     
     // Add top panel to main
     mainPanel.addNorth(topPanel, 30.0);
     // Add Product name
-    topPanel.add(new ui.Html("<div id=\"header-content\"><strong>Dart Web Toolkit Showcase</strong> (ver 0.2)</div>"));
+    //topPanel.add(new ui.Html("<div id=\"header-content\"><strong>Dart Web Toolkit Showcase</strong> (ver 0.2)</div>"));
+    ui.Label headerContent = new ui.Label("Dart Web Toolkit Showcase (ver 0.2)");
+    headerContent.addStyleName("header-content");
+    topPanel.add(headerContent);
     
     // Create status content
     ui.DockLayoutPanel statusPanel = new ui.DockLayoutPanel(util.Unit.PX);
@@ -60,7 +52,9 @@ class AppView extends ui.Composite implements AppPresenterDisplay {
     // Add top panel to main
     mainPanel.addSouth(statusPanel, 20.0);
     // Add Product name
-    ui.Html copyright = new ui.Html("<div style=\"text-align:right;\">Copyright 2012-2013 Sergey Akopkokhyants&nbsp;</div>");
+    //ui.Html copyright = new ui.Html("<div style=\"text-align:right;\">Copyright 2012-2013 Sergey Akopkokhyants&nbsp;</div>");
+    ui.Label copyright = new ui.Label("Copyright 2012-2013 Sergey Akopkokhyants");
+    copyright.addStyleName("copyright");
     statusPanel.add(copyright);
 
     // Create Center panel